[DEBT-043-D11] Split portfolio/pnl.py into focused modules
- Split pnl.py (554 LoC) into 4 files:
  - _pnl_models.py (80): PnLSummary, Lot, EffectiveTrade, FifoResult
  - _fifo.py (174): Trade normalization and FIFO calculation
  - _settlements.py (191): Settlement price/fee parsing, synthesis
  - pnl.py (276): PnLCalculator facade with backwards-compat interface
- All files now ≤400 lines (target met)
- Public API preserved via class attribute aliases
- Quality gates passed: pre-commit, pytest (1052 passed)
- Acceptance criteria: D11/D18 complete
